Амангелді_Айдос_Задание_14/requirements.txt
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
fastapi
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
asyncpg
pytest
pytest-asyncio
aioredis
redis
celery

